Refactor `test_python_frontend.py` (#2701)
`test_python_frontend.py` is now 4k+ lines which is tedious to work
with.

This PR:
1. Moves `exec_nvfuser` into a separate `class NVFuserTest (TestCase)`
which can be used as the base class of any nvfuser test class that needs
to verify serialization/deserialization.
2. Removes `setUpModule` into `setup_class` of `NVFuserTest` to avoid
adding it in every file. This method will run `atexit_serde_check` to
enable default serialization at the end of the test suite.
~3. As a starting point, moving matmul and linear tests into a separate
file.~ To be done in separate PR.
4. Replaces `unittest` with `pytest` for consistency.
